+++ FlutterWave.Core/Models/Services/Foundations/ExternalFlutterWave/ExternalTransfers/ExternalTransferRatesResponse.cs
@@ -31,7 +31,7 @@ namespace FlutterWave.Core.Models.Services.Foundations.ExternalFlutterWave.Exter
             public string Currency { get; set; }
 
             [JsonProperty("amount")]
-            public int Amount { get; set; }
+            public decimal Amount { get; set; }
         }
 
 
@@ -44,7 +44,7 @@ namespace FlutterWave.Core.Models.Services.Foundations.ExternalFlutterWave.Exter
             public string Currency { get; set; }
 
             [JsonProperty("amount")]
-            public int Amount { get; set; }
+            public decimal Amount { get; set; }
         }
