+++ FlutterWave.Core/Models/Services/Foundations/FlutterWave/Misc/BalanceByCurrencyResponse.cs
@@ -9,7 +9,7 @@
         public class BalanceCurrencyData
         {
             public string Currency { get; set; }
-            public int AvailableBalance { get; set; }
+            public decimal AvailableBalance { get; set; }
             public double LedgerBalance { get; set; }
         }
